Move common config types to mod.rs
